DBZ-4459: Introduce interfaces and default implementations for change event source metrics
